Cambiando tipos de datos del dto

+++ app/src/main/java/com/example/criptoapp/ui/screens/components/CoinItem.kt
@@ -47,7 +47,7 @@ fun CoinItem(
             }
             val decimalFormat = DecimalFormat("#,###.######")
             Text(
-                text = "$ " + decimalFormat.format(coin.valor.toDouble()),
+                text = "$ " + decimalFormat.format(coin.valor),
                 textAlign = TextAlign.End,
                 fontWeight = FontWeight.Normal,
                 letterSpacing = 3.sp,
